Added the final Check-in File
The SWF file is an output from the .fla file. It can be used in a
webpage to navigate through instructions on how to retrieve i-94, scan
immigration documents and use the iHawk system to upload check-in
documents to ISS.

You are allowed to choose your category as F-1, J-1 or other to find
correct appropriate instructions for your immigration status.